move ja3 implementation dir to ztls, few improvements
